[Madhu/Shivaji]#228962 - Add adobe launch script for BYOR analytics
