+++ src/index.html
@@ -58,10 +58,11 @@
           <a href="https://www.thoughtworks.com/radar/how-to-byor" target="_blank">Google Sheet, CSV or JSON</a> file
           below…
         </p>
-        <form method="get">
+        <form id='byor-form-submit' method="get">
           <input
             type="text"
             name="documentId"
+            id ="document-input"
             placeholder="e.g. https://docs.google.com/spreadsheets/d/<sheetid> or hosted CSV/JSON file"
             required=""
           />
