+++ src/site.js
@@ -1,7 +1,7 @@
 require('./common')
 require('./images/logo.png')
 require('./images/radar_legend.png')
-require('./gtm.js')
+require('./analytics.js')
 
 const Factory = require('./util/factory')
